refactor: move styles to your own file
