+++ components/Carousel/styles.ts
@@ -1,0 +1,25 @@
+import styled from 'styled-components';
+import { motion } from 'framer-motion';
+
+export const Wrapper = styled(motion.div)`
+  max-width: 100%;
+  height: 10rem;
+  display: flex;
+  align-items: center;
+  overflow-x: hidden;
+`;
+
+export const Carousel = styled(motion.ul)`
+  display: flex;
+  cursor: grab;
+`;
+
+export const Slide = styled(motion.li)`
+  position: relative;
+  margin-right: 0.5rem;
+
+  .template {
+    border-radius: 4px;
+    pointer-events: none;
+  }
+`;
