+++ components/Carousel/types.ts
@@ -1,0 +1,6 @@
+import { Template } from 'shared/apiSchema';
+
+export type CarouselProps = {
+  templates: Template[];
+  onSeletedTemplate: (template: Template) => void;
+};
