+++ components/Layout/Container/types.ts
@@ -1,0 +1,6 @@
+export type ContainerVariants = 'auto' | 'full' | 'transparent';
+
+export type ContainerProps = {
+  background?: string;
+  variant?: ContainerVariants;
+};
