+++ components/Layout/Header/index.tsx
@@ -1,7 +1,7 @@
 import { LocaleSwitcher } from 'components/LocaleSwitcher';
 import { ToggleTheme } from 'components/ToggleTheme';
 
-import * as S from 'styles/components/Layout/Header';
+import * as S from './styles';
 
 export function Header() {
   return (
