+++ components/LocaleSwitcher/index.tsx
@@ -4,7 +4,7 @@ import useSound from 'use-sound';
 
 import menuOpenSound from 'public/sounds/menu-open.mp3';
 
-import * as S from 'styles/components/LocaleSwitcher';
+import * as S from './styles';
 
 export function LocaleSwitcher() {
   const router = useRouter();
