+++ components/LocaleSwitcher/styles.ts
