--- components/Shimmer/types.ts
@@ -1,4 +1,0 @@
-export type ShimmerProps = {
-  w: string | number;
-  h: string | number;
-};
