+++ components/Sparkles/Sparkle/types.ts
@@ -1,0 +1,8 @@
+export type SparkleProps = {
+  color: string;
+  size: number;
+  style: {
+    top: string;
+    left: string;
+  };
+};
