+++ components/Sparkles/styles.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.strong`
+  position: relative;
+  z-index: 1;
+  font-weight: bold;
+`;
