+++ components/ToastMessage/types.ts
@@ -1,0 +1,8 @@
+import { CSSProperties } from 'react';
+
+import { ToastMessageData } from 'shared/toasts';
+
+export type ToastMessageProps = {
+  message: ToastMessageData;
+  style?: CSSProperties;
+};
