+++ hooks/useCaptionMeme/types.ts
@@ -1,0 +1,5 @@
+export type CaptionMeme = {
+  data: {
+    url: string;
+  };
+};
