+++ hooks/useMemes/index.ts
@@ -1,9 +1,17 @@
 import { useQuery, UseQueryOptions } from 'react-query';
 
-import { getMemes } from 'lib/imgflip';
+import { api } from 'config/client';
 import { Template } from 'shared/apiSchema';
 
+import { GetMemes } from './types';
+
 export const GET_MEMES_QUERY_KEY = 'memes';
+
+export async function getMemes() {
+  return api
+    .get<GetMemes>('/get_memes')
+    .then(response => response.data.data.memes);
+}
 
 export function useMemes(options?: UseQueryOptions<Template[]>) {
   return useQuery<Template[]>(GET_MEMES_QUERY_KEY, getMemes, options);
