+++ hooks/useMemes/types.ts
@@ -1,0 +1,7 @@
+import { Template } from 'shared/apiSchema';
+
+export type GetMemes = {
+  data: {
+    memes: Template[];
+  };
+};
