+++ hooks/useRandomInterval/types.ts
@@ -1,0 +1,5 @@
+export type Props = {
+  callback: () => void;
+  minDelay: number;
+  maxDelay: number;
+};
