+++ layouts/Home/types.ts
@@ -1,0 +1,9 @@
+import { Template } from 'shared/apiSchema';
+
+export type HomeProps = {
+  templates: Template[];
+};
+
+export type Box = Template['box_count'];
+
+export type DownloadFile = Pick<Template, 'url' | 'name'>;
